add highlight

+++ pkg/meta/v2/query_dsl.go
@@ -270,8 +270,8 @@ type AggregationDateHistogram struct {
 }
 
 type Highlight struct {
-	NumberOfFragments int64                `json:"number_of_fragments"`
-	FragmentSize      int64                `json:"fragment_size"`
+	NumberOfFragments int                  `json:"number_of_fragments"`
+	FragmentSize      int                  `json:"fragment_size"`
 	PreTags           []string             `json:"pre_tags"`
 	PostTags          []string             `json:"post_tags"`
 	Fields            map[string]Highlight `json:"fields"`
